+++ indico/modules/users/controllers.py
@@ -54,8 +54,8 @@ from indico.modules.users.schemas import (AffiliationSchema, BasicCategorySchema
 from indico.modules.users.util import (anonymize_user, get_avatar_url_from_name, get_gravatar_for_user,
                                        get_linked_events, get_mastodon_server_name, get_related_categories,
                                        get_suggested_categories, get_unlisted_events, get_user_by_email,
-                                       get_user_titles, merge_users, search_users, send_avatar, serialize_user,
-                                       set_user_avatar)
+                                       get_user_titles, merge_users, search_affiliations, search_users, send_avatar,
+                                       serialize_user, set_user_avatar)
 from indico.modules.users.views import (WPUser, WPUserDashboard, WPUserDataExport, WPUserFavorites, WPUserPersonalData,
                                         WPUserProfilePic, WPUsersAdmin)
 from indico.util.date_time import now_utc
@@ -310,18 +310,7 @@ class RHPersonalDataUpdate(RHUserBase):
 class RHSearchAffiliations(RH):
     @use_kwargs({'q': fields.String(load_default='')}, location='query')
     def _process(self, q):
-        exact_match = db.func.lower(Affiliation.name) == q.lower()
-        q_filter = Affiliation.name.ilike(f'%{q}%') if len(q) > 2 else exact_match
-        res = (
-            Affiliation.query
-            .filter(~Affiliation.is_deleted, q_filter)
-            .order_by(
-                exact_match.desc(),
-                db.func.lower(Affiliation.name).startswith(q.lower()).desc(),
-                db.func.lower(Affiliation.name)
-            )
-            .limit(20)
-            .all())
+        res = search_affiliations(q)
         return AffiliationSchema(many=True).jsonify(res)